requirements.txt
alembic==1.8.1
Babel==2.10.3
bcrypt==3.2.2
blinker==1.5
cffi==1.15.1
click==8.1.3
Flask==2.2.1
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.2
Flask-Mail==0.9.1
Flask-Migrate==3.1.0
Flask-Principal==0.4.0
Flask-Script==2.0.6
Flask-Security==3.0.0
Flask-SQLAlchemy==2.5.1
Flask-Testing==0.8.1
Flask-WTF==1.0.1
greenlet==1.1.2
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
Mako==1.2.1
MarkupSafe==2.1.1
mysqlclient==2.1.1
passlib==1.7.4
pycparser==2.21
pytz==2022.1
PyYAML==6.0
speaklater==1.3
SQLAlchemy==1.4.39
Werkzeug==2.2.1
WTForms==3.0.1
zipp==3.8.1
